refactor(client-utils): rename module
