+++ packages/libs/client-utils/src/coin/get-balance.ts
@@ -2,7 +2,7 @@ import type { ChainId } from '@kadena/client';
 import { Pact } from '@kadena/client';
 import { execution } from '@kadena/client/fp';
 
-import { dirtyReadClient } from '../core/rich-client';
+import { dirtyReadClient } from '../core/client-helpers';
 import type { IClientConfig } from '../core/utils/helpers';
 
 import { pipe } from 'ramda';
