+++ packages/libs/client-utils/src/core/index.ts
@@ -1,2 +1,2 @@
-export * from './rich-client';
+export * from './client-helpers';
 export * from './utils/asyncPipe';
